// src/app/shared/_components/search/models/search.model.ts
export class Search {
  SearchId: Number|null;
  CustomersId: Number|null;
  UnitOwnersId: Number|null;
  PropertyCondition: Number|null;
  Furniture: Number|null;
  DateSearch: Date|null;
  EvacuationDate: Date|null;
  PublishDate: Date|null;
  City: string|null;
  Neighborhood: string|null;
  Street: string|null;
  Description: string|null;
  Number?: Number|null;
  FromFloor?: Number|null;
  ToFloor?: Number|null;
  FromArea?: Number|null;
  ToArea?: Number|null;
  MinRoomsNum: Number|null;
  MaxRoomsNum: Number|null;
  MinPrice: Number|null;
  MaxPrice: Number|null;
  ViewsNum: Number|null;
  Relevant: Boolean|null;
  Parking: Boolean|null;
  PandorDoors: Boolean|null;
  SolarHeater: Boolean|null;
  AirConditioning: Boolean|null;
  AccessForDisabled: Boolean|null;
  Animal: Boolean|null;
  SecurityRoom: Boolean|null;
  Elevator: Boolean|null;
  Terrace: Boolean|null;

  constructor() {
    this.SearchId = 0;
    this.CustomersId = 1;
    // this.UnitOwnersId = 0;
    // this.PropertyCondition = 0;
    // this.Furniture = 0;
    this.DateSearch = new Date();
    // this.EvacuationDate = new Date();
    // this.PublishDate = new Date();
    // this.City = '';
    // this.Neighborhood = '';
    // this.Street = '';
    // this.Description = '';
    // this.Number = 0;
    // this.FromFloor = 0;
    // this.ToFloor = 0;
    // this.FromArea = 0;
    // this.ToArea = 0;
    // this.MinRoomsNum = 0;
    // this.MaxRoomsNum = 0;
    // this.MinPrice = 0;
    // this.MaxPrice = 0;
    // this.ViewsNum = 0;
    // this.Relevant = false;
    // this.Parking = false;
    // this.PandorDoors = false;
    // this.SolarHeater = false;
    // this.AirConditioning = false;
    // this.AccessForDisabled = false;
    // this.Animal = false;
    // this.SecurityRoom = false;
    // this.Elevator = false;
    // this.Terrace = false;
  }
}